+++ NEWS.md
@@ -40,6 +40,7 @@
 - Fixed issue where new R package projects did not inherit "Generate documentation with Roxygen" preference
 - Fixed an issue where large character vectors were shown with an NaN size in the environment pane (#15919)
 - Fixed an issue where hitting the Escape key to close the "Update Available" dialog would exit RStudio (#15444)
+- Fixed an issue where the splash screen would not close and the RStudio main window would not show when starting RStudio Desktop (#16191)
 
 #### Posit Workbench